Call window refactor, split into seperate components & composables
