+++ app/src/components/call/widgets/TranscriberWidget.vue
